backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
sqlparse==0.5.3
tzdata==2024.2


annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.58.1
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
openai-whisper
pyttsx3
#torch
#TTS==0.15.1
#textgrid

#torch==2.0.1


whitenoise
TTS
#g2p-en
#torch
#python-Levenshtein
#torchaudio

django-cors-headers


azure-cognitiveservices-speech
requests
#celery[redis]
#celery[beat]
# Указание репозитория для CPU-версий
#torch==2.0.1+cpu
#torchaudio==2.0.1+cpu
#--extra-index-url https://download.pytorch.org/whl/cpu
#TTS==0.15.1
#TTS==0.25.1

